+++ src/main/java/com/springboot/socialhub_api/api/repositories/FollowingsRepository.java
@@ -3,5 +3,5 @@ package com.springboot.socialhub_api.api.repositories;
 import com.springboot.socialhub_api.api.model.Followings;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FollowingsRepository extends JpaRepository<Followings,Long> {
+public interface FollowingsRepository extends JpaRepository<Followings,Integer> {
 }
